requirements.txt
Flask==2.3.0
openpyxl==3.0.10
azure-cosmos==4.0.0